change price from float to decimal

+++ migrations/sqls/20220421220017-products-up.sql
@@ -1,6 +1,6 @@
 CREATE TABLE products (
    id SERIAL PRIMARY KEY,
    name VARCHAR(50) NOT NULL,
-   price FLOAT(3) NOT NULL,
+   price DECIMAL NOT NULL,
    stock INTEGER NOT NULL 
 );
